cuda : use 64-bit loop counter in argmax_f32

+++ ggml/src/ggml-cuda/argmax.cu
@@ -24,11 +24,11 @@ static __global__ void argmax_f32(const float * __restrict__ x, int32_t * __rest
     int   argmax = -1;
     const float * rowx = x + row * ncols;
 
-    for (int32_t col = threadIdx.x; col < ncols; col += blockDim.x) {
+    for (int64_t col = threadIdx.x; col < ncols; col += blockDim.x) {
         const float val = rowx[col];
         if (val > maxval) {
             maxval = val;
-            argmax = col;
+            argmax = (int) col;
         }
     }
